+++ test/Test-Automation/src/PP_TestUserGroups.xml
@@ -13,10 +13,10 @@
 				<class name="Execution.homePageTest"/>
 			</classes>
 		</parameter>
-	</test>   
+	</test>    
 
   
-	 
+ -->	
    		
 	<test name="User Access Groups">
 		<parameter name="env" value="https://localhost/LEAF_Request_Portal/admin/?a=mod_groups">
@@ -27,7 +27,8 @@
 		</parameter>
 	</test>
 
--->
+
+<!--   
 
 	<test name="User Access Groups TESTING">
 		<parameter name="env" value="https://localhost/LEAF_Request_Portal/admin/?a=mod_groups">
@@ -38,8 +39,19 @@
 		</parameter>
 	</test>
 
+ --> 
 
- <!--
+	<test name="User Access Groups Part 2">
+		<parameter name="env" value="https://localhost/LEAF_Request_Portal/admin/">
+		<parameter name="browser" value="chrome"/>
+			<classes>
+				<class name="Execution.userAccessGroupsPart2"/>
+			</classes>
+		</parameter>
+	</test>
+
+
+ <!--  
 	
 		
 	<test name="Admin Page">
