+++ test/Test-Automation/test-output/RegressionSuite/User Access Groups Part 2.html
@@ -1,0 +1,119 @@
+<html>
+<head>
+<title>TestNG:  User Access Groups Part 2</title>
+<link href="../testng.css" rel="stylesheet" type="text/css" />
+<link href="../my-testng.css" rel="stylesheet" type="text/css" />
+
+<style type="text/css">
+.log { display: none;} 
+.stack-trace { display: none;} 
+</style>
+<script type="text/javascript">
+<!--
+function flip(e) {
+  current = e.style.display;
+  if (current == 'block') {
+    e.style.display = 'none';
+    return 0;
+  }
+  else {
+    e.style.display = 'block';
+    return 1;
+  }
+}
+
+function toggleBox(szDivId, elem, msg1, msg2)
+{
+  var res = -1;  if (document.getElementById) {
+    res = flip(document.getElementById(szDivId));
+  }
+  else if (document.all) {
+    // this is the way old msie versions work
+    res = flip(document.all[szDivId]);
+  }
+  if(elem) {
+    if(res == 0) elem.innerHTML = msg1; else elem.innerHTML = msg2;
+  }
+
+}
+
+function toggleAllBoxes() {
+  if (document.getElementsByTagName) {
+    d = document.getElementsByTagName('div');
+    for (i = 0; i < d.length; i++) {
+      if (d[i].className == 'log') {
+        flip(d[i]);
+      }
+    }
+  }
+}
+
+// -->
+</script>
+
+</head>
+<body>
+<h2 align='center'>User Access Groups Part 2</h2><table border='1' align="center">
+<tr>
+<td>Tests passed/Failed/Skipped:</td><td>8/0/0</td>
+</tr><tr>
+<td>Started on:</td><td>Tue Sep 12 14:03:47 CDT 2023</td>
+</tr>
+<tr><td>Total time:</td><td>16 seconds (16200 ms)</td>
+</tr><tr>
+<td>Included groups:</td><td></td>
+</tr><tr>
+<td>Excluded groups:</td><td></td>
+</tr>
+</table><p/>
+<small><i>(Hover the method name to see the test class name)</i></small><p/>
+<table width='100%' border='1' class='invocation-passed'>
+<tr><td colspan='4' align='center'><b>PASSED TESTS</b></td></tr>
+<tr><td><b>Test method</b></td>
+<td width="30%"><b>Exception</b></td>
+<td width="10%"><b>Time (seconds)</b></td>
+<td><b>Instance</b></td>
+</tr>
+<tr>
+<td title='Execution.userAccessGroupsPart2.navigateAdminBack()'><b>navigateAdminBack</b><br>Test class: Execution.userAccessGroupsPart2</td>
+<td></td>
+<td>0</td>
+<td>Execution.userAccessGroupsPart2@5562c41e</td></tr>
+<tr>
+<td title='Execution.userAccessGroupsPart2.scrollUp()'><b>scrollUp</b><br>Test class: Execution.userAccessGroupsPart2</td>
+<td></td>
+<td>1</td>
+<td>Execution.userAccessGroupsPart2@5562c41e</td></tr>
+<tr>
+<td title='Execution.userAccessGroupsPart2.syncServices()'><b>syncServices</b><br>Test class: Execution.userAccessGroupsPart2</td>
+<td></td>
+<td>1</td>
+<td>Execution.userAccessGroupsPart2@5562c41e</td></tr>
+<tr>
+<td title='Execution.userAccessGroupsPart2.openAccessGroup()'><b>openAccessGroup</b><br>Test class: Execution.userAccessGroupsPart2</td>
+<td></td>
+<td>1</td>
+<td>Execution.userAccessGroupsPart2@5562c41e</td></tr>
+<tr>
+<td title='Execution.userAccessGroupsPart2.gotoAdminPanel()'><b>gotoAdminPanel</b><br>Test class: Execution.userAccessGroupsPart2</td>
+<td></td>
+<td>3</td>
+<td>Execution.userAccessGroupsPart2@5562c41e</td></tr>
+<tr>
+<td title='Execution.userAccessGroupsPart2.scrollDown()'><b>scrollDown</b><br>Test class: Execution.userAccessGroupsPart2</td>
+<td></td>
+<td>2</td>
+<td>Execution.userAccessGroupsPart2@5562c41e</td></tr>
+<tr>
+<td title='Execution.userAccessGroupsPart2.openUserAccessGroups()'><b>openUserAccessGroups</b><br>Test class: Execution.userAccessGroupsPart2</td>
+<td></td>
+<td>2</td>
+<td>Execution.userAccessGroupsPart2@5562c41e</td></tr>
+<tr>
+<td title='Execution.userAccessGroupsPart2.testForCertPage()'><b>testForCertPage</b><br>Test class: Execution.userAccessGroupsPart2</td>
+<td></td>
+<td>1</td>
+<td>Execution.userAccessGroupsPart2@5562c41e</td></tr>
+</table><p>
+</body>
+</html>
