+++ test/Test-Automation/test-output/RegressionSuite/User Access Groups Part 2.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!-- Generated by org.testng.reporters.JUnitXMLReporter -->
+<testsuite ignored="0" hostname="DESKTOP-OOHKJP3" failures="0" tests="8" name="User Access Groups Part 2" time="16.2" errors="0" timestamp="2023-09-12T14:04:03 CDT">
+  <testcase classname="Execution.userAccessGroupsPart2" name="testForCertPage" time="1.477"/>
+  <testcase classname="Execution.userAccessGroupsPart2" name="scrollDown" time="2.018"/>
+  <testcase classname="Execution.userAccessGroupsPart2" name="syncServices" time="1.997"/>
+  <testcase classname="Execution.userAccessGroupsPart2" name="gotoAdminPanel" time="3.455"/>
+  <testcase classname="Execution.userAccessGroupsPart2" name="navigateAdminBack" time="0.333"/>
+  <testcase classname="Execution.userAccessGroupsPart2" name="openUserAccessGroups" time="2.014"/>
+  <testcase classname="Execution.userAccessGroupsPart2" name="scrollUp" time="1.014"/>
+  <testcase classname="Execution.userAccessGroupsPart2" name="openAccessGroup" time="1.426"/>
+</testsuite> <!-- User Access Groups Part 2 -->
