+++ test/Test-Automation/test-output/junitreports/TEST-Execution.userAccessGroupsPart2.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!-- Generated by org.testng.reporters.JUnitReportReporter -->
+<testsuite hostname="DESKTOP-OOHKJP3" failures="0" tests="8" name="Execution.userAccessGroupsPart2" time="13.737" errors="0" timestamp="2023-09-12T14:04:03 CDT" skipped="0">
+  <testcase classname="Execution.userAccessGroupsPart2" name="testForCertPage" time="1.477"/>
+  <system-out/>
+  <testcase classname="Execution.userAccessGroupsPart2" name="scrollDown" time="2.019"/>
+  <system-out/>
+  <testcase classname="Execution.userAccessGroupsPart2" name="syncServices" time="1.998"/>
+  <system-out/>
+  <testcase classname="Execution.userAccessGroupsPart2" name="gotoAdminPanel" time="3.456"/>
+  <system-out/>
+  <testcase classname="Execution.userAccessGroupsPart2" name="navigateAdminBack" time="0.333"/>
+  <system-out/>
+  <testcase classname="Execution.userAccessGroupsPart2" name="openUserAccessGroups" time="2.014"/>
+  <system-out/>
+  <testcase classname="Execution.userAccessGroupsPart2" name="scrollUp" time="1.014"/>
+  <system-out/>
+  <testcase classname="Execution.userAccessGroupsPart2" name="openAccessGroup" time="1.426"/>
+  <system-out/>
+</testsuite> <!-- Execution.userAccessGroupsPart2 -->
