+++ test/Test-Automation/test-output/old/RegressionSuite/testng.xml.html
@@ -1,1 +1,1 @@
-<html><head><title>testng.xml for RegressionSuite</title></head><body><tt>&lt;?xml&nbsp;version="1.0"&nbsp;encoding="UTF-8"?&gt;<br/>&lt;!DOCTYPE&nbsp;suite&nbsp;SYSTEM&nbsp;"https://testng.org/testng-1.0.dtd"&gt;<br/>&lt;suite&nbsp;name="RegressionSuite"&nbsp;guice-stage="DEVELOPMENT"&gt;<br/>&nbsp;&nbsp;&lt;test&nbsp;thread-count="5"&nbsp;name="User&nbsp;Access&nbsp;Groups&nbsp;TESTING"&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&lt;parameter&nbsp;name="browser"&nbsp;value="chrome"/&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&lt;parameter&nbsp;name="env"&nbsp;value="https://localhost/LEAF_Request_Portal/admin/?a=mod_groups"/&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&lt;classes&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&lt;class&nbsp;name="Execution.userAccessGroupsTESTING"/&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&lt;/classes&gt;<br/>&nbsp;&nbsp;&lt;/test&gt;&nbsp;&lt;!--&nbsp;User&nbsp;Access&nbsp;Groups&nbsp;TESTING&nbsp;--&gt;<br/>&lt;/suite&gt;&nbsp;&lt;!--&nbsp;RegressionSuite&nbsp;--&gt;<br/></tt></body></html>
+<html><head><title>testng.xml for RegressionSuite</title></head><body><tt>&lt;?xml&nbsp;version="1.0"&nbsp;encoding="UTF-8"?&gt;<br/>&lt;!DOCTYPE&nbsp;suite&nbsp;SYSTEM&nbsp;"https://testng.org/testng-1.0.dtd"&gt;<br/>&lt;suite&nbsp;name="RegressionSuite"&nbsp;guice-stage="DEVELOPMENT"&gt;<br/>&nbsp;&nbsp;&lt;test&nbsp;thread-count="5"&nbsp;name="User&nbsp;Access&nbsp;Groups"&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&lt;parameter&nbsp;name="browser"&nbsp;value="chrome"/&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&lt;parameter&nbsp;name="env"&nbsp;value="https://localhost/LEAF_Request_Portal/admin/?a=mod_groups"/&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&lt;classes&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&lt;class&nbsp;name="Execution.userAccessGroups"/&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&lt;/classes&gt;<br/>&nbsp;&nbsp;&lt;/test&gt;&nbsp;&lt;!--&nbsp;User&nbsp;Access&nbsp;Groups&nbsp;--&gt;<br/>&nbsp;&nbsp;&lt;test&nbsp;thread-count="5"&nbsp;name="User&nbsp;Access&nbsp;Groups&nbsp;Part&nbsp;2"&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&lt;parameter&nbsp;name="browser"&nbsp;value="chrome"/&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&lt;parameter&nbsp;name="env"&nbsp;value="https://localhost/LEAF_Request_Portal/admin/"/&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&lt;classes&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&lt;class&nbsp;name="Execution.userAccessGroupsPart2"/&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&lt;/classes&gt;<br/>&nbsp;&nbsp;&lt;/test&gt;&nbsp;&lt;!--&nbsp;User&nbsp;Access&nbsp;Groups&nbsp;Part&nbsp;2&nbsp;--&gt;<br/>&lt;/suite&gt;&nbsp;&lt;!--&nbsp;RegressionSuite&nbsp;--&gt;<br/></tt></body></html>
