+++ test/Test-Automation/test-output/old/RegressionSuite/toc.html
@@ -8,9 +8,9 @@
 <h3><p align="center">Results for<br/><em>RegressionSuite</em></p></h3>
 <table border='1' width='100%'>
 <tr valign='top'>
-<td>1 test</td>
-<td><a target='mainFrame' href='classes.html'>1 class</a></td>
-<td>26 methods:<br/>
+<td>2 tests</td>
+<td><a target='mainFrame' href='classes.html'>2 classes</a></td>
+<td>77 methods:<br/>
 &nbsp;&nbsp;<a target='mainFrame' href='methods.html'>chronological</a><br/>
 &nbsp;&nbsp;<a target='mainFrame' href='methods-alphabetical.html'>alphabetical</a><br/>
 &nbsp;&nbsp;<a target='mainFrame' href='methods-not-run.html'>not run (0)</a></td>
@@ -22,8 +22,16 @@
 </tr></table>
 <table width='100%' class='test-failed'>
 <tr><td>
-<table style='width: 100%'><tr><td valign='top'>User Access Groups TESTING (23/3/0)</td><td valign='top' align='right'>
-  <a href='User Access Groups TESTING.html' target='mainFrame'>Results</a>
+<table style='width: 100%'><tr><td valign='top'>User Access Groups (56/13/0)</td><td valign='top' align='right'>
+  <a href='User Access Groups.html' target='mainFrame'>Results</a>
+</td></tr></table>
+</td></tr><p/>
+</table>
+
+<table width='100%' class='test-passed'>
+<tr><td>
+<table style='width: 100%'><tr><td valign='top'>User Access Groups Part 2 (8/0/0)</td><td valign='top' align='right'>
+  <a href='User Access Groups Part 2.html' target='mainFrame'>Results</a>
 </td></tr></table>
 </td></tr><p/>
 </table>
