+++ test/Test-Automation/test-output/old/index.html
@@ -4,6 +4,6 @@
 </head><body>
 <h2><p align='center'>Test results</p></h2>
 <table border='1' width='100%' class='main-page'><tr><th>Suite</th><th>Passed</th><th>Failed</th><th>Skipped</th><th>testng.xml</th></tr>
-<tr align='center' class='invocation-failed'><td><em>Total</em></td><td><em>23</em></td><td><em>3</em></td><td><em>0</em></td><td>&nbsp;</td></tr>
+<tr align='center' class='invocation-failed'><td><em>Total</em></td><td><em>64</em></td><td><em>13</em></td><td><em>0</em></td><td>&nbsp;</td></tr>
 <tr align='center' class='invocation-failed'><td><a href='RegressionSuite/index.html'>RegressionSuite</a></td>
-<td>23</td><td>3</td><td>0</td><td><a href='RegressionSuite/testng.xml.html'>Link</a></td></tr></table></body></html>
+<td>64</td><td>13</td><td>0</td><td><a href='RegressionSuite/testng.xml.html'>Link</a></td></tr></table></body></html>
