+++ test/Test-Automation/test-output/testng-failed.xml
@@ -1,20 +1,30 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd">
 <suite name="Failed suite [RegressionSuite]" guice-stage="DEVELOPMENT">
-  <test thread-count="5" name="User Access Groups TESTING(failed)">
+  <test thread-count="5" name="User Access Groups(failed)">
     <parameter name="browser" value="chrome"/>
     <parameter name="env" value="https://localhost/LEAF_Request_Portal/admin/?a=mod_groups"/>
     <classes>
-      <class name="Execution.userAccessGroupsTESTING">
+      <class name="Execution.userAccessGroups">
         <methods>
           <include name="setUp"/>
           <include name="initializeFramework"/>
-          <include name="closeDown"/>
-          <include name="clickAddUserInNexus2"/>
+          <include name="clickAssignBackup2"/>
           <include name="clickAddUserInNexus3"/>
-          <include name="inputNexusEmployee2"/>
+          <include name="openAccessGroup6"/>
+          <include name="clickAssignBackup"/>
+          <include name="inputEmployee"/>
+          <include name="confirmYes3"/>
+          <include name="navigateNexusBack"/>
+          <include name="saveEmployee"/>
+          <include name="selectMemberPosition2"/>
+          <include name="selectMemberPosition1"/>
+          <include name="openAccessGroup"/>
+          <include name="closeDown"/>
+          <include name="addNexusUser_2"/>
+          <include name="inputNexusEmployee3"/>
         </methods>
-      </class> <!-- Execution.userAccessGroupsTESTING -->
+      </class> <!-- Execution.userAccessGroups -->
     </classes>
-  </test> <!-- User Access Groups TESTING(failed) -->
+  </test> <!-- User Access Groups(failed) -->
 </suite> <!-- Failed suite [RegressionSuite] -->
